keymaster: realign android service with trusty
All the logics/code in the keymaster service could be exactly the same
as trusty, there is nothing specific for OP-TEE.
Compared to trusty the main difference are with OP-TEE:
- forward operations from IPC to the Trusted Application
- support crypto operations in the Trusted Application

Thus all the code related to the keymaster service is now realigned
with trusty: system/core/trusty/keymaster
We follow the exact same tree structure and naming convention.
All occurrences of "trusty" word are replaced by "optee".

The update is based on this HEAD reference in trusty:
f53626de0 Merge cherrypicks of ['googleplex-android-review.googlesource.com/25156721'] into udc-release.

In that way all the futur fixes/improvments made in trusty service can
be applied easily for OP-TEE keymaster.

All the crypto operations not yet supported should be reported by the
Trusted Application.

Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Signed-off-by: Julien Masson <jmasson@baylibre.com>